requirements.txt
django
djangorestframework
django-cors-headers
pillow
user_agents
pyotp
requests